+++ app/encoding.go
@@ -3,10 +3,10 @@ package app
 import (
 	"github.com/cosmos/cosmos-sdk/std"
 
-	"github.com/persistenceOne/pstake-native/v4/app/params"
-	liquidstakeibctypes "github.com/persistenceOne/pstake-native/v4/x/liquidstakeibc/types"
-	lscosmostypes "github.com/persistenceOne/pstake-native/v4/x/lscosmos/types"
-	ratesynctypes "github.com/persistenceOne/pstake-native/v4/x/ratesync/types"
+	"github.com/persistenceOne/pstake-native/v5/app/params"
+	liquidstakeibctypes "github.com/persistenceOne/pstake-native/v5/x/liquidstakeibc/types"
+	lscosmostypes "github.com/persistenceOne/pstake-native/v5/x/lscosmos/types"
+	ratesynctypes "github.com/persistenceOne/pstake-native/v5/x/ratesync/types"
 )
 
 // MakeEncodingConfig creates an EncodingConfig for testing