+++ x/liquidstake/abci.go
@@ -7,8 +7,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/persistenceOne/pstake-native/v4/x/liquidstake/keeper"
-	"github.com/persistenceOne/pstake-native/v4/x/liquidstake/types"
+	"github.com/persistenceOne/pstake-native/v5/x/liquidstake/keeper"
+	"github.com/persistenceOne/pstake-native/v5/x/liquidstake/types"
 )
 
 func BeginBlock(ctx context.Context, k keeper.Keeper) {