+++ x/liquidstake/keeper/keeper.go
@@ -9,8 +9,9 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	mintkeeper "github.com/cosmos/cosmos-sdk/x/mint/keeper"
+	liquidkeeper "github.com/cosmos/gaia/v24/x/liquid/keeper"
 
-	"github.com/persistenceOne/pstake-native/v4/x/liquidstake/types"
+	"github.com/persistenceOne/pstake-native/v5/x/liquidstake/types"
 )
 
 // Keeper of the liquidstake store
@@ -24,6 +25,7 @@ type Keeper struct {
 	mintKeeper     mintkeeper.Keeper
 	distrKeeper    types.DistrKeeper
 	slashingKeeper types.SlashingKeeper
+	liquidKeeper   liquidkeeper.Keeper
 
 	router    *baseapp.MsgServiceRouter
 	authority string
@@ -39,6 +41,7 @@ func NewKeeper(
 	mintKeeper mintkeeper.Keeper,
 	distrKeeper types.DistrKeeper,
 	slashingKeeper types.SlashingKeeper,
+	liquidKeeper liquidkeeper.Keeper,
 	router *baseapp.MsgServiceRouter,
 	authority string,
 ) Keeper {
@@ -56,6 +59,7 @@ func NewKeeper(
 		mintKeeper:     mintKeeper,
 		distrKeeper:    distrKeeper,
 		slashingKeeper: slashingKeeper,
+		liquidKeeper:   liquidKeeper,
 		router:         router,
 		authority:      authority,
 	}