+++ x/liquidstake/module_test.go
@@ -10,9 +10,9 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
-	chain "github.com/persistenceOne/pstake-native/v4/app"
-	"github.com/persistenceOne/pstake-native/v4/x/liquidstake"
-	"github.com/persistenceOne/pstake-native/v4/x/liquidstake/types"
+	chain "github.com/persistenceOne/pstake-native/v5/app"
+	"github.com/persistenceOne/pstake-native/v5/x/liquidstake"
+	"github.com/persistenceOne/pstake-native/v5/x/liquidstake/types"
 )
 
 type ModuleTestSuite struct {