+++ x/liquidstake/types/rebalancing_test.go
@@ -6,7 +6,7 @@ import (
 	"cosmossdk.io/math"
 	"github.com/stretchr/testify/require"
 
-	"github.com/persistenceOne/pstake-native/v4/x/liquidstake/types"
+	"github.com/persistenceOne/pstake-native/v5/x/liquidstake/types"
 )
 
 var liquidValidators = []types.LiquidValidator{